+++ epigenetics/README.md
@@ -6,25 +6,35 @@ Especially targeted to functions related to the analysis of epigenomics data. It
 
 _in ./chipseq.py_
 
-- dropWeirdChromosomes: 
-- extractPairedSingleEndFrom: 
-- findReplicates: 
-- computeSingleEnd: 
-- computePairedEnd: 
-- bigWigFrom: 
-- mergeBams: 
-- loadPeaks: 
-- pysam_computePeaksAt: 
-- bedtools_computePeaksAt: 
-- computePeaksAt: 
-- simpleMergedPeaks: 
-- findpeakpath: 
-- mergeReplicatePeaks: 
-- findAdditionalPeaks: 
-- putInConscensus: 
-- findAdditionalCobindingSignal: 
-- fullDiffPeak: 
-- diffPeak: 
+- bigWigFrom
+- ReadRoseSuperEnhancers
+- loadPeaks
+- pysam_getPeaksAt
+- bedtools_getPeaksAt
+- makeProfiles
+- getPeaksAt
+- computeMeanCov
+- substractPeaks
+- simpleMergePeaks
+- findpeakpath
+- findBestPeak
+- mergeReplicatePeaks
+- col_nan_scatter
+- col_nan_kde_histo
+- findAdditionalPeaks
+- negLogLikelihood
+- poissonFit
+- putInBed
+- pairwiseOverlap
+- enrichment
+- findAdditionalCobindingSignal
+- fullDiffPeak
+- diffPeak
+- MakeSuperEnhancers
+- runChromHMM
+- loadMEMEmotifs
+- simpleMergeMotifs
+- substractPeaksTo
 
 ## highly recommended packages
