+++ sequencing/README.md
@@ -4,8 +4,14 @@ A set of function to help work with sequencing data (bed files, bam files, fastq
 
 ## Contains
 
-fromGTF2BED: transforms a GTF file to a BED file, only works for some GTFs for now
+- fromGTF2BED: transforms a GTF file to a BED file, only works for some GTFs for now
 getBamDate: parses a bam file header to try to compute when it was generated (as best as it can, if it has had many modification done to it across a long span of time, you will receive the average of that)
+- dropWeirdChromosomes
+- extractPairedSingleEndFrom
+- findReplicates
+- singleEnd
+- pairedEnd
+- mergeBams
 
 ## Other very recommended tools
