+++ utils/helper.py
@@ -445,3 +445,15 @@ def compareDfs(df1, df2):
     print('FOUND new NAs in df1: ' + str(newNAs))
     print('FOUND new 0s in df1: ' + str(new0s))
     return nmissmatchCols, omissmatchCols, nmissmatchInds, omissmatchInds, newNAs, new0s
+
+def stringifydict(res):
+    """
+    
+    """
+    a = {}
+    for k,v in res.items():
+        if type(v) is dict:
+            a[k] = stringifydict(v)
+        else:
+            a[str(k)] = v
+    return a
